fix comment controller selesai

+++ app/Http/Controllers/CommentController.php
@@ -11,12 +11,11 @@ class CommentController extends Controller
     function store(Request $request){
         $request->validate([
             'article_id' => 'required',
-            'user_id' => 'required',
             'comment' => 'required|max:255',
         ]);
 
         $commentData = [
-            'user_id' => $request->user_id,
+            'user_id' => $request->user()->id,
             'article_id' => $request->article_id,
             'comment' => $request->comment
         ];
@@ -38,6 +37,16 @@ class CommentController extends Controller
             'updated' => true
         ];
 
+        $userId = Comment::where('id', $request->id)->first()->user_id;
+
+        if ($userId != $request->user()->id) {
+            return ResponseFormatter::response(
+                404,
+                'error',
+                'anda tidak terautentikasi'
+            );
+        }
+
         Comment::where('id', $request->id)->update($commentData);
 
         return ResponseFormatter::response(200, 'success', $commentData);
@@ -50,7 +59,18 @@ class CommentController extends Controller
             return ResponseFormatter::response(404, 'error', "comment tidak ada");
         }
 
+        $userId = Comment::where('id', $request->id)->first()->user_id;
+
+        if ($userId != $request->user()->id) {
+            return ResponseFormatter::response(
+                404,
+                'error',
+                'anda tidak terautentikasi'
+            );
+        }
+
         Comment::where('id', $request->id)->delete();
+
         return ResponseFormatter::response(200, 'success', $comment);
     }
 }
